+++ CHANGELOG.md
@@ -6,8 +6,11 @@
 
 - some more MetBrewer palettes
 - some more ColorBrewer palettes
+- dependency on ColorVectorSpaces
 
 ## Changed
+
+- optimized `get()` - #91 - thanks @stevengj !
 
 ### Removed